+++ serialization/COM3D2/arc/arc_hashtable.go
@@ -12,19 +12,22 @@ import (
 	"github.com/MeidoPromotionAssociation/MeidoSerialization/serialization/binaryio/stream"
 )
 
+// makeCountedSliceForAppend 按线格式计数创建受限的初始切片容量，避免损坏计数触发巨额预分配
+// makeCountedSliceForAppend creates a capped initial slice capacity from a wire count to prevent excessive preallocation from corrupt counts
 func makeCountedSliceForAppend[T any](count int32) []T {
 	if count <= 0 {
 		return make([]T, 0)
 	}
 	const maxInitialCapacity = 1024
-	capacity := int(count)
+	capacity := count
 	if capacity > maxInitialCapacity {
 		capacity = maxInitialCapacity
 	}
 	return make([]T, 0, capacity)
 }
 
-// readHashTable reads a hash table from a binary stream and returns a pointer to the constructed hashTable or an error.
+// readHashTable 从二进制流读取哈希表，并返回构建的 hashTable 或错误
+// readHashTable reads a hash table from a binary stream and returns a pointer to the constructed hashTable or an error
 func readHashTable(reader *stream.BinaryReader) (*hashTable, error) {
 	var ht hashTable
 	header, err := reader.ReadInt64()
@@ -73,7 +76,7 @@ func readHashTable(reader *stream.BinaryReader) (*hashTable, error) {
 	ht.Padding = padding
 
 	ht.DirEntries = makeCountedSliceForAppend[fileEntryRec](ht.DirCount)
-	for i := 0; i < int(ht.DirCount); i++ {
+	for i := int32(0); i < ht.DirCount; i++ {
 		var e fileEntryRec
 		hash, err := reader.ReadUInt64()
 		if err != nil {
@@ -89,7 +92,7 @@ func readHashTable(reader *stream.BinaryReader) (*hashTable, error) {
 	}
 
 	ht.FileEntries = makeCountedSliceForAppend[fileEntryRec](ht.FileCount)
-	for i := 0; i < int(ht.FileCount); i++ {
+	for i := int32(0); i < ht.FileCount; i++ {
 		var e fileEntryRec
 		hash, err := reader.ReadUInt64()
 		if err != nil {
@@ -104,7 +107,7 @@ func readHashTable(reader *stream.BinaryReader) (*hashTable, error) {
 		ht.FileEntries = append(ht.FileEntries, e)
 	}
 	ht.ParentsID = makeCountedSliceForAppend[uint64](ht.Depth)
-	for i := 0; i < int(ht.Depth); i++ {
+	for i := int32(0); i < ht.Depth; i++ {
 		parentHash, err := reader.ReadUInt64()
 		if err != nil {
 			return nil, fmt.Errorf("read parent hash failed: %w", err)
@@ -112,7 +115,7 @@ func readHashTable(reader *stream.BinaryReader) (*hashTable, error) {
 		ht.ParentsID = append(ht.ParentsID, parentHash)
 	}
 	ht.SubDirEntries = makeCountedSliceForAppend[*hashTable](ht.DirCount)
-	for i := 0; i < int(ht.DirCount); i++ {
+	for i := int32(0); i < ht.DirCount; i++ {
 		subDir, err := readHashTable(reader)
 		if err != nil {
 			return nil, fmt.Errorf("read subDir entry failed: %w", err)
@@ -122,8 +125,10 @@ func readHashTable(reader *stream.BinaryReader) (*hashTable, error) {
 	return &ht, nil
 }
 
-// readNameTable reads a table of names from a binary stream and returns a map of hashes to strings.
-// It stops reading when the end of the stream is reached or an error occurs.
+// readNameTable 从二进制流读取名称表并返回哈希到字符串的映射
+// 到达流末尾或发生错误时停止读取
+// readNameTable reads a table of names from a binary stream and returns a map of hashes to strings
+// It stops reading when the end of the stream is reached or an error occurs
 func readNameTable(reader *stream.BinaryReader) (map[uint64]string, error) {
 	lut := make(map[uint64]string)
 	for {
@@ -143,12 +148,13 @@ func readNameTable(reader *stream.BinaryReader) (map[uint64]string, error) {
 			return nil, fmt.Errorf("invalid name size")
 		}
 
-		utf16leString, err := reader.ReadBytes(int(nameSize) * 2)
+		utf16leString, err := reader.ReadBytes(int64(nameSize) * 2)
 		if err != nil {
 			return nil, fmt.Errorf("read name bytes failed: %w", err)
 		}
 
-		// UTF-16LE to string
+		// 将 UTF-16LE 转换为字符串
+		// Convert UTF-16LE to a string
 		name := utf16leToString(utf16leString)
 		if _, exists := lut[nameHash]; !exists {
 			lut[nameHash] = name
@@ -157,13 +163,14 @@ func readNameTable(reader *stream.BinaryReader) (map[uint64]string, error) {
 	return lut, nil
 }
 
-// writeHashTable writes the hash table for the current directory and its subdirectories to the provided BinaryWriter.
+// writeHashTable 将当前目录及其子目录的哈希表写入 BinaryWriter
+// writeHashTable writes the hash table for the current directory and its subdirectories to the provided BinaryWriter
 func (arc *Arc) writeHashTable(bw *stream.BinaryWriter, dirOffsets map[uint64]int64, uuidToHash map[uint64]uint64, fileOffsets map[uint64]int64, cur *Dir) error {
-	dirCount, err := checkedArcInt32Count("directory count", len(cur.Dirs))
+	dirCount, err := checkedArcInt32Count("directory count", int64(len(cur.Dirs)))
 	if err != nil {
 		return err
 	}
-	fileCount, err := checkedArcInt32Count("file count", len(cur.Files))
+	fileCount, err := checkedArcInt32Count("file count", int64(len(cur.Files)))
 	if err != nil {
 		return err
 	}
@@ -195,6 +202,7 @@ func (arc *Arc) writeHashTable(bw *stream.BinaryWriter, dirOffsets map[uint64]in
 		return fmt.Errorf("write padding failed: %w", err)
 	}
 
+	// 目录条目按 dirOffsets 排序
 	// Directory entries ordered by dirOffsets
 	dirs := cur.sortedDirs()
 	sort.Slice(dirs, func(i, j int) bool { return dirOffsets[dirs[i].UniqueID()] < dirOffsets[dirs[j].UniqueID()] })
@@ -207,6 +215,7 @@ func (arc *Arc) writeHashTable(bw *stream.BinaryWriter, dirOffsets map[uint64]in
 		}
 	}
 
+	// 文件条目按 uuidToHash 升序排序
 	// File entries ordered by uuidToHash ascending
 	files := cur.sortedFiles()
 	sort.Slice(files, func(i, j int) bool { return uuidToHash[files[i].UniqueID()] < uuidToHash[files[j].UniqueID()] })
@@ -219,21 +228,25 @@ func (arc *Arc) writeHashTable(bw *stream.BinaryWriter, dirOffsets map[uint64]in
 		}
 	}
 
+	// 父哈希从父目录到根目录收集后反向写入
+	// 收集父目录
 	// Parent hashes from parent up to root reversed
-	// collect parents
+	// Collect parents
 	var parents []uint64
 	p := cur.Parent
 	for p != nil {
 		parents = append(parents, uuidToHash[p.UniqueID()])
 		p = p.Parent
 	}
-	// write reversed
-	for i := len(parents) - 1; i >= 0; i-- {
+	// 反向写入
+	// Write reversed
+	for i := int64(len(parents)) - 1; i >= 0; i-- {
 		if err := bw.WriteUInt64(parents[i]); err != nil {
 			return fmt.Errorf("write parent hash failed: %w", err)
 		}
 	}
 
+	// 子表
 	// Subtables
 	for _, d := range dirs {
 		if err := arc.writeHashTable(bw, dirOffsets, uuidToHash, fileOffsets, d); err != nil {
@@ -243,9 +256,11 @@ func (arc *Arc) writeHashTable(bw *stream.BinaryWriter, dirOffsets map[uint64]in
 	return nil
 }
 
-// writeNameTable writes the name table, including names, hashes, and their UTF-16LE encoded byte size, to the provided BinaryWriter.
+// writeNameTable 将名称、哈希和 UTF-16LE 字节大小写入名称表
+// writeNameTable writes the name table, including names, hashes, and their UTF-16LE encoded byte size, to the provided BinaryWriter
 func (arc *Arc) writeNameTable(bw *stream.BinaryWriter, utf16 bool) error {
-	// gather files, dirs, and root, distinct by name, preserving order for determinism
+	// 收集文件、目录和根目录名称，按名称去重并保留顺序以保证确定性
+	// Gather files, dirs, and root, distinct by name, preserving order for determinism
 	var names []string
 	seen := make(map[string]bool)
 	add := func(n string) {
@@ -255,6 +270,7 @@ func (arc *Arc) writeNameTable(bw *stream.BinaryWriter, utf16 bool) error {
 		}
 	}
 
+	// 遵循 C# 顺序：FileCount、DirCount、Root
 	// Follow C# order: FileCount, then DirCount, then Root
 	for _, f := range AllFiles(arc) {
 		add(f.Name)
@@ -264,18 +280,21 @@ func (arc *Arc) writeNameTable(bw *stream.BinaryWriter, utf16 bool) error {
 	}
 	add(arc.Root.Name)
 
-	// write pairs
+	// 写出名称与哈希对
+	// Write pairs
 	for _, n := range names {
 		var h uint64
+		// C# 中的 Bytes 和 Size 始终使用 UTF-16LE 字节与字符数量
+		// 只有 Hash 取决于 utf16 参数
 		// In C#, Bytes and Size are always UTF-16LE and character count
-		// only the Hash depends on the utf16 parameter
+		// Only the Hash depends on the utf16 parameter
 		if utf16 {
 			h = NameHashUTF16(n)
 		} else {
 			h = NameHashUTF8(n)
 		}
 		b := utf16le(n)
-		sz, err := checkedArcInt32Count(fmt.Sprintf("name %q UTF-16 code-unit count", n), len(b)/2)
+		sz, err := checkedArcInt32Count(fmt.Sprintf("name %q UTF-16 code-unit count", n), int64(len(b))/2)
 		if err != nil {
 			return err
 		}
@@ -293,24 +312,29 @@ func (arc *Arc) writeNameTable(bw *stream.BinaryWriter, utf16 bool) error {
 	return nil
 }
 
-// lastIndexOfSep returns the index of the last occurrence of the file path separator in the given string.
-// If the separator is not found, it returns -1.
-func lastIndexOfSep(s string) int {
+// lastIndexOfSep 返回给定字符串中最后一个路径分隔符的索引
+// 未找到分隔符时返回 -1
+// lastIndexOfSep returns the index of the last occurrence of the file path separator in the given string
+// If the separator is not found, it returns -1
+func lastIndexOfSep(s string) int64 {
 	sep := string(filepath.Separator)
-	return bytes.LastIndex([]byte(s), []byte(sep))
+	return int64(bytes.LastIndex([]byte(s), []byte(sep)))
 }
 
-// utf16leToString converts a UTF-16LE encoded byte slice into a string (utf-8), truncating one trailing byte if the length is odd.
+// utf16leToString 将 UTF-16LE 字节切片转换为 UTF-8 字符串，长度为奇数时截断末尾字节
+// utf16leToString converts a UTF-16LE encoded byte slice into a UTF-8 string, truncating one trailing byte if the length is odd
 func utf16leToString(b []byte) string {
 	if len(b)%2 != 0 {
 		b = b[:len(b)-1]
 	}
-	// decode pairs into runes
+	// 将字节对解码为 rune
+	// Decode pairs into runes
 	u16 := make([]uint16, len(b)/2)
 	for i := 0; i < len(u16); i++ {
 		u16[i] = binary.LittleEndian.Uint16(b[i*2:])
 	}
-	// convert to runes
+	// 转换为 rune
+	// Convert to runes
 	runes := utf16.Decode(u16)
 	return string(runes)
 }
